New start for the project
Changed some data around for convenience's sake.

As for now I would like to channel the different color values so the script can detect an array of colors and apply the future shape detection to it. Why? Because with the original Binary setup we were only able to retrieve shadows of shapes that passed the threshold provided. This leaves us with irregular shapes at times. If we separate shape detection via colors then an art like the one I provided should be sectioned off by color and its respective shapes

Later on in the project we can switch back to Binary detection as I think it will still be better, but that is when we figure out shape shadow slicing to retrieve the appropriate shapes.